+++ weldx/welding/__init__.py
@@ -1,0 +1,1 @@
+from . import iso_9692_1
